Added: new validations to google map component

+++ Resources/views/frontend/components/map/google.blade.php
@@ -1,5 +1,7 @@
 <script>
 
+    console.log("Isite::Components|Map|Google|Init")
+
     //Global Variables
     let map;
     let bounds;
@@ -13,26 +15,30 @@
     @endif
 
     //Validation to Label Location
-    @if(isset($locationName) && !is_null($locationName))
+    @if(isset($showLocationName) && $showLocationName)
       labelLocation = "{{$locationName}}"
     @endif
 
-    //Validation Modal
-    @if(!isset($inModal) || !$inModal)
-      document.addEventListener("DOMContentLoaded", function () {
+    //Se le agrego esta validacion de "usingLivewire", xq cuando se le daba al boton "Agregar nueva direccion" ya no mostraba el mapa
+    @if($usingLivewire==false)
+      //Validation Modal
+      @if(!isset($inModal) || !$inModal)
+        document.addEventListener("DOMContentLoaded", function () {
+      @endif
     @endif
 
     //Validation Geocoder to get address
     @if($allowMoveMarker)
-      geocoder = new google.maps.Geocoder();
+      var geocoder = new google.maps.Geocoder();
     @endif
-
-
+     
     /*
     * Google Map | INIT
     */
     function initMap() 
     {   
+      console.log("Isite::Components|Map|Google|initMap")
+
         const position = {lat: {{$lat}}, lng: {{$lng}}};
        
         map = new google.maps.Map(document.getElementById("{{$mapId}}"), {
@@ -45,36 +51,38 @@
           setSimpleMarkerGoogle(position)
         @endif
 
-        //Allow User Move a marker | Prueba 1
+        
+        //Multiple Locations By Default | Al segundo intento no funciona
         /*
-        @if($allowMoveMarker)
-          google.maps.event.addListener(map, 'click', function(event) {
-            deleteMarkersGoogle()
-            setSimpleMarkerGoogle(event.latLng)
-            var clickData = event.latLng.toJSON();
-
-            console.warn(event)
-            //console.warn(clickData)
-            // Emit to Component: Iprofile | Address Form
-            //window.livewire.emit('selectedMarkerInMap',clickData);
-          });
+        @if(!is_null($locations) && count($locations))
+          //console.warn("Default Multiple Locations")
+          @foreach($locations as $location)
+            let locationPosition = {lat: {{$location['lat']}}, lng: {{$location['lng']}}};
+            //console.warn("Position: "+locationPosition)
+            setSimpleMarkerGoogle(locationPosition)
+          @endforeach
         @endif
         */
+        
+        
        
-
+        //return newMap
     }
 
     //INIT GOOGLE MAPS
     initMap();
 
-    //Validation Modal
-    @if(!isset($inModal) || !$inModal) 
-      });
+    //Se le agrego esta validacion de "usingLivewire", xq cuando se le daba al boton "Agregar nueva direccion" ya no mostraba el mapa
+    @if($usingLivewire==false)
+      //Validation Modal
+      @if(!isset($inModal) || !$inModal) 
+        });
+      @endif
     @endif
 
 
     /*
-    * Set Simple mark
+    * Set Simple mark | First Time
     */
     function setSimpleMarkerGoogle(position)
     {
@@ -82,7 +90,7 @@
           position: position, 
           map: map,
           @if(!is_null($imageIcon)) icon: urlMarkerIcon, @endif
-          @if(!is_null($locationName))  label: labelLocation, @endif
+          @if(!is_null($showLocationName))  label: labelLocation, @endif
           @if($allowMoveMarker) draggable:true @endif
       });
 
@@ -105,7 +113,8 @@
     */
     function setMarkerGoogle(bounds)
     {
-      
+      console.warn('SET MARKER GOOGLE') 
+
       const itemPosition = {lat: parseFloat(locLat), lng: parseFloat(locLng)};
       //const iconMarker = "https://dev-lip.ozonohosting.com/assets/media/marker-2.png?u=1705935415";
     
@@ -163,6 +172,35 @@
     }
 
     /*
+    * Get Data Especific Address Google
+    */
+    function getDataFromAddressGoogle(addressComponents)
+    {
+        var components = {};
+        var city;
+        var postalCode = null;
+        var state;
+        var country;
+      
+        jQuery.each(addressComponents, function(k,v1) {
+            jQuery.each(v1.types,function(k2,v2){
+                components[v2] = v1.short_name
+            });
+        });
+                
+        if(components.locality) { city = components.locality; }
+        if(!city) { city = components.administrative_area_level_1; }
+        if(components.postal_code) { postalCode = components.postal_code; }
+        if(components.administrative_area_level_1) { state = components.administrative_area_level_1; }
+        if(components.country) { country = components.country; }
+        //console.warn("COUNTRY: "+country+" | STATE: "+state+" | CITY: "+city+" | Postal: "+postalCode)
+        //Data Final
+        var addressData = {country: country, state: state, city: city, postal: postalCode};
+
+        return addressData;
+    }
+
+    /*
     * Get Address Format 
     */
     function getAddressFromPosition(pos) 
@@ -176,7 +214,7 @@
           addressFormat = responses[0].formatted_address;
 
           //method in  = Address Form | Auto Complete Google
-          var addressData = getDataFromAddress(responses[0].address_components)
+          var addressData = getDataFromAddressGoogle(responses[0].address_components)
 
         } else {
           addressFormat = 'Cannot determine address at this location.';
@@ -204,25 +242,27 @@
     }
 
     /*
-    * LIVEWIRE | Listener Component | Case Address Form
+    * LIVEWIRE | Listener Component | Case Address Form | Update Marker with Google Autocomplete
     */
-    window.addEventListener('google-update-marker-in-map', event => {
-      //Before delete markers
-      deleteMarkersGoogle()
+    @if(isset($allowMoveMarker))
+      window.addEventListener('google-update-marker-in-map', event => {
+        //Before delete markers
+        deleteMarkersGoogle()
 
-      //Init and Reset Bounds
-      var bounds = new google.maps.LatLngBounds();
+        //Init and Reset Bounds
+        var bounds = new google.maps.LatLngBounds();
 
-      //Get data
-      itemPosition = event.detail.itemPosition
+        //Get data
+        var itemPosition = event.detail.itemPosition
 
-      //Create Marker
-      setSimpleMarkerGoogle(itemPosition)
+        //Create Marker
+        setSimpleMarkerGoogle(itemPosition)
 
-      //Set Bounds
-      bounds.extend(itemPosition);
-      map.fitBounds(bounds);
+        //Set Bounds
+        bounds.extend(itemPosition);
+        map.fitBounds(bounds);
 
-    });
+      });
+    @endif
   
 </script>
